+++ enclave/src/lib.rs
@@ -55,10 +55,18 @@ use std::string::String;
 use std::vec::Vec;
 
 use utils::*;
+use rsa3072::*;
+use ed25519::*;
+use state::*;
+use crate::aes::*;
 
 mod constants;
 mod utils;
 mod attestation;
+mod rsa3072;
+mod ed25519;
+mod state;
+mod aes;
 
 pub mod cert;
 pub mod hex;
